Ignore ssh-agent for dotcloud connections
The ssh client tries keys provided by ssh-agent before falling back to the
dotcloud key specified on the command-line with -i. However this can
ultimately result in authentication failure (breaking dotcloud run and
push) if ssh-agent has six or more keys cached, due to the default value of
MaxAuthTries 6 in sshd.

This can be avoided using the IdentitiesOnly configuration option to ssh to
ignore ssh-agent.